+++ src/App/Options/ThemeOption.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace WpTheme\Scaffold\App\Options;
+
+use WpTheme\Scaffold\Core\Theme;
+
+use WpTheme\Scaffold\Core\Abstracts\Option;
+
+use WpTheme\Scaffold\Core\Contracts\OptionInterface; 
+
+final class ThemeOption extends Option implements OptionInterface {
+
+  public function getGroup ( string $group ): array {
+
+    $found = array_filter( $this->getOption(), function ( $key ) use ( $group ) {
+
+      return false !== strpos( $key, "$group." );
+    }, ARRAY_FILTER_USE_KEY );
+
+    return \wp_parse_args( $found, array_filter( $this->getDefault(), function ( $key ) use ( $group ) {
+
+      return false !== strpos( $key, "$group." );
+    }, ARRAY_FILTER_USE_KEY ) );
+  }
+
+  public static function register ( Theme $theme ) {
+
+    return new Self( 
+      'theme_option',
+      'edit_themes',
+      $theme->make( \WpTheme\Scaffold\Core\Resources\Storage::class, [
+        'default' => array_merge( $theme['theme.styles'], $theme['theme.supports'] )
+      ])
+    );
+  }
+}
